requirements.txt
langchain>=0.1.11
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
streamlit>=1.28.1
python-dotenv>=1.0.0
PyPDF2>=3.0.1
pydantic>=2.5.0
requests>=2.31.0
huggingface-hub>=0.19.0
transformers>=4.35.0
torch>=2.0.0
sentence-transformers>=2.2.0
